refactor(ui.animation): consolidate animation helpers into existing files
Drop ui/color.go, ui/palette.go and ui/fade.go. The whole animation
surface is small enough that the extra files added navigation cost
without earning much
